Revert "Merge branch 'develop'"
This reverts commit 756fb973c9e82c00d3cbd94cc4c2542d84d3e186, reversing
changes made to d54bfc8d5e382bb497c97ac1743a3e721dc6811f.
