+++ tomato_method_001.xcodeproj/xcuserdata/macalexander.xcuserdatad/xcschemes/xcschememanagement.plist
@@ -4,27 +4,6 @@
 <dict>
 	<key>SchemeUserState</key>
 	<dict>
-		<key>SnapKitPlayground (Playground) 1.xcscheme</key>
-		<dict>
-			<key>isShown</key>
-			<false/>
-			<key>orderHint</key>
-			<integer>2</integer>
-		</dict>
-		<key>SnapKitPlayground (Playground) 2.xcscheme</key>
-		<dict>
-			<key>isShown</key>
-			<false/>
-			<key>orderHint</key>
-			<integer>3</integer>
-		</dict>
-		<key>SnapKitPlayground (Playground).xcscheme</key>
-		<dict>
-			<key>isShown</key>
-			<false/>
-			<key>orderHint</key>
-			<integer>1</integer>
-		</dict>
 		<key>tomato_method_001.xcscheme_^#shared#^_</key>
 		<dict>
 			<key>orderHint</key>
