+++ tomato_method_001/Assets.xcassets/AppIcon.appiconset/Contents.json
@@ -1,7 +1,6 @@
 {
   "images" : [
     {
-      "filename" : "111.jpeg",
       "idiom" : "universal",
       "platform" : "ios",
       "size" : "1024x1024"
