--- tomato_method_001/Sources/Extensions/UITextField+ Extensions.swift
@@ -1,26 +1,0 @@
-
-import UIKit
-
-// MARK: - Extensions
-
-extension UITextField {
-    func setLeftIcon(_ image: UIImage) {
-        let iconView = UIImageView(frame: CGRect(x: 20, y: 5, width: 20, height: 20))
-        iconView.image = image
-        let iconContainerView: UIView = UIView(frame: CGRect(x: 20, y: 0, width: 30, height: 30))
-        iconContainerView.addSubview(iconView)
-        leftView = iconContainerView
-        leftViewMode = .always
-    }
-}
-
-extension UITextField {
-    func setRightIcon(_ image: UIImage) {
-        let iconView = UIImageView(frame: CGRect(x: -5, y: 5, width: 20, height: 20))
-        iconView.image = image
-        let iconContainerView: UIView = UIView(frame: CGRect(x: 0, y: 0, width: 30, height: 30))
-        iconContainerView.addSubview(iconView)
-        rightView = iconContainerView
-        rightViewMode = .always
-    }
-}
