--- tomato_method_001/Sources/UIViewControllers/CircularBarProgressView.swift
@@ -1,54 +1,0 @@
-
-/*
- import UIKit
-
-class CircularProgressBarView: UIView {
-    
-    override init(frame: CGRect) {
-        super.init(frame: frame)
-    }
-    
-    required init?(coder: NSCoder) {
-        super.init(coder: coder)
-    }
-    
-    private var circleLayer = CAShapeLayer()
-    private var progressLayer = CAShapeLayer()
-    
-    private var startAngle = CGFloat(-Double.pi / 2)
-    private var endAngle = CGFloat(3 * Double.pi / 2)
-    
-    func createCircularPath() {
-        let circularPath = UIBezierPath(arcCenter: CGPoint(x: frame.size.width / 2.0, y: frame.size.height / 2.0), radius: 80, startAngle: startAngle, endAngle: endAngle, clockwise: true)
-        
-        circleLayer.path = circularPath.cgPath
-        
-        circleLayer.fillColor = UIColor.clear.cgColor
-        circleLayer.lineCap = .round
-        circleLayer.lineWidth = 20.0
-        circleLayer.strokeEnd = 1.0
-        circleLayer.strokeColor = UIColor.white.cgColor
-        layer.addSublayer(circleLayer)
-        
-        progressLayer.path = circularPath.cgPath
-        
-        circleLayer.fillColor = UIColor.clear.cgColor
-        circleLayer.lineCap = .round
-        circleLayer.lineWidth = 10.0
-        circleLayer.strokeEnd = 0
-        circleLayer.strokeColor = UIColor.red.cgColor
-        layer.addSublayer(progressLayer)
-    }
-    
-    func progressAnimation(duration: TimeInterval) {
-        let circularProgressAnimation = CABasicAnimation(keyPath: "strokeEnd")
-        
-        circularProgressAnimation.duration = duration
-        circularProgressAnimation.toValue = 1.0
-        circularProgressAnimation.fillMode = .forwards
-        circularProgressAnimation.isRemovedOnCompletion = false
-        progressLayer.add(circularProgressAnimation, forKey: "progressAnim")
-    }
-}
-*/
-
