+++ tomato_method_001/ViewController.swift
@@ -1,0 +1,19 @@
+//
+//  ViewController.swift
+//  tomato_method_001
+//
+//  Created by Mac Alexander on 13.11.2023.
+//
+
+import UIKit
+
+class ViewController: UIViewController {
+
+    override func viewDidLoad() {
+        super.viewDidLoad()
+        // Do any additional setup after loading the view.
+    }
+
+
+}
+
